moved signaluser model to signalsapp and migrate data and adjust history view
