+++ api/app/signals/apps/signals/factories/signal_user.py
@@ -1,7 +1,7 @@
 from django.contrib.auth import get_user_model
 from factory import DjangoModelFactory, Sequence, SubFactory, post_generation
 
-from signals.apps.users.models import SignalUser
+from signals.apps.signals.models.signal_user import SignalUser
 
 User = get_user_model()
